+++ tests/qml/imports/qs/Commons/Border.qml
@@ -2,6 +2,7 @@ pragma Singleton
 import QtQuick
 
 QtObject {
+  function none() { return {} }
   function flat(color, width) {
     return { color: color, width: width }
   }