+++ tests/qml/imports/qs/Commons/Util.qml
@@ -1,0 +1,6 @@
+pragma Singleton
+import QtQuick
+
+QtObject {
+  function alpha(color, amount) { return Qt.rgba(color.r, color.g, color.b, amount) }
+}