+++ tests/qml/imports/qs/Ui/BorderSurface.qml
@@ -2,4 +2,5 @@ import QtQuick
 
 Rectangle {
   property var borderSpec: ({})
+  property real borderRight: 0
 }